requirements.txt
Flask
pandas
scikit-learn
nltk
spacy
pdfminer.six
python-docx
gunicorn
ddgs
googlesearch-python
